Revert moving dns record expiry into a component file
